Cleanup done. Class Port split. Many temporaries.
