+++ src/main/resources/de/michab/scream/extensions/i-test.s
@@ -1,0 +1,5 @@
+
+(define i-test 8)
+
+(load "i-test2.s")
+
