+++ src/main/resources/de/michab/scream/extensions/i-test2.s
@@ -1,0 +1,3 @@
+
+
+
